refactor: implement code quality improvements (linting, coverage, JSDoc)
