+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 dist
 .serena
 
+coverage/
