+++ package.json
@@ -11,10 +11,13 @@
   "scripts": {
     "test": "node --test tests/**/*.test.js",
     "lint": "eslint . --max-warnings=0",
+    "coverage": "c8 --reporter=lcov --reporter=text npm test",
     "start:stdio": "node src/cli.js",
     "start:remote": "node src/cli.js --mode remote --port 3333 --cors *",
     "dev:stdio": "node --watch src/cli.js",
-    "dev:remote": "node --watch src/cli.js --mode remote --port 3333 --cors *"
+    "dev:remote": "node --watch src/cli.js --mode remote --port 3333 --cors *",
+    "duplicate-check": "jscpd src",
+    "unused-deps": "depcheck"
   },
   "engines": {
     "node": ">=18"
@@ -23,6 +26,13 @@
     "@modelcontextprotocol/sdk": "^0.5.0"
   },
   "devDependencies": {
-    "eslint": "^8.57.1"
+    "@eslint/js": "^9.36.0",
+    "c8": "^10.1.3",
+    "c8": "^9.1.0",
+    "depcheck": "^1.4.7",
+    "eslint": "^8.57.1",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "globals": "^16.4.0",
+    "jscpd": "^4.0.5"
   }
 }
