+++ src/server/register-transcript-tool.js
@@ -1,4 +1,5 @@
 import { transcriptYt } from '../tool/transcriptYt.js'
+import { transcriptSchema } from '../lib/transcriptSchema.js'
 
 /**
  * Registers the transcript_yt MCP tool on the provided MCP server instance.
@@ -9,19 +10,7 @@ export function registerTranscriptTool(server) {
   server.registerTool({
     name: 'transcript_yt',
     description: 'Fetches YouTube transcript segments from a video URL for LLM consumption.',
-    inputSchema: {
-      type: 'object',
-      properties: {
-        videoUrl: { type: 'string', description: 'Full YouTube video URL.' },
-        preferredLanguages: {
-          type: 'array',
-          items: { type: 'string' },
-          description: "Optional ordered language codes preference, e.g., ['pt-BR','en']."
-        }
-      },
-      required: ['videoUrl'],
-      additionalProperties: false
-    }
+   inputSchema: transcriptSchema
   }, async (args = {}) => {
     const videoUrl = String(args.videoUrl || '')
     const preferredLanguages = Array.isArray(args.preferredLanguages) ? args.preferredLanguages : undefined
