+++ src/server/sdk-config.js
@@ -1,19 +1,9 @@
 import { SSEServerTransport } from '@modelcontextprotocol/sdk/server/sse.js'
 import { createServerConfig } from './config.js'
 import { McpServer } from './mcp-server.js'
+import { safeImportStreamableTransport } from '../lib/safeImport.js'
 
-let StreamableHTTPServerTransport = null
-try {
-  const module = await import('@modelcontextprotocol/sdk/server/streamableHttp.js')
-  if (module && module.StreamableHTTPServerTransport) {
-    StreamableHTTPServerTransport = module.StreamableHTTPServerTransport
-  }
-} catch (error) {
-  const message = typeof error === 'object' && error !== null ? String(error.message || error) : String(error)
-  if (error?.code !== 'ERR_MODULE_NOT_FOUND' && !message.includes('Cannot find module')) {
-    throw error
-  }
-}
+const StreamableHTTPServerTransport = await safeImportStreamableTransport()
 
 export function createSdkServerConfig(overrides = {}) {
   const baseConfig = createServerConfig({ ...overrides, mode: 'remote' })
